libapp: move core of ScrollComponent to Base::ScrollComponent
This is done to prepare to support scrolling in libtui.
